\name{CEU}
\alias{CEU}
\title{
Sampled SNPs from all SNPs of CEU population
}
\description{
A GRange object \code{CEU} contains 5\% of all SNPs from CEU by
controling genome-wide density is the same as all SNPs from CEU
}

\usage{
data(CEU)
}

\value{
	The data frame \code{CEU} contains three columns,
	\item{SNP}{SNP rs number}
	\item{seqnames}{Chromosome number associated with rs number}
	\item{ranges}{Chromosomal position, in base pairs, associated with rs number}
}

\author{
Li Chen <li.chen@emory.edu>, Zhaohui S.Qin<zhaohui.qin@emory.edu>
}



