\name{Tcell}
\alias{Tcell}
\title{
Peak regions of H3K4me1 in Peripheral blood T cell
}
\description{
A GRange object \code{Tcell} contains three columns: chromosome, genomic start position and genomic end position.}

\usage{
data(Tcell)
}

\value{
	The data frame \code{Tcell} contains three columns,
	\item{seqnames}{Chromosome id}
	\item{ranges}{Chromosome position}
}

\author{
Li Chen <li.chen@emory.edu>, Zhaohui S.Qin<zhaohui.qin@emory.edu>
}

