\name{traseR-package}
\alias{traseR-package}
\docType{package}
\title{
GWAS trait-associated SNP enrichment analyses in genomic intervals
}
\description{
Perform GWAS trait-associated SNP enrichment analyses in genomic intervals. Explore and visualize the results.
}
\details{
\tabular{ll}{
Package: \tab traseR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-08\cr
License: \tab GPL \cr
}
}
\author{
Li Chen <li.chen@emory.edu>, Zhaohui S.Qin<zhaohui.qin@emory.edu>
}

\keyword{ package }
